ddl: move some dist-reorg code to dist_* files
